Split HydrogenFunction into source and header
