+++ src/analyticfunctions/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(mrchem PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/HydrogenFunction.cpp
+    )
